added Suspense and sceletons, updated README
