+++ app/(ui)/dashboard/(overview)/loading.tsx
@@ -1,0 +1,5 @@
+import DashboardSkeleton from '@/app/(ui)/skeletons';
+ 
+export default function Loading() {
+  return <DashboardSkeleton />;
+}
