+++ app/(ui)/dashboard/customers/page.tsx
