+++ app/(ui)/dashboard/customers/table.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
-import { lusitana } from '@/app/ui/fonts';
-import Search from '@/app/ui/search';
+import { lusitana } from '@/app/(ui)/fonts';
+import Search from '@/app/(ui)/search';
 import { CustomersTable, FormattedCustomersTable } from '@/app/lib/definitions';
 
 export default async function CustomersTable({
