+++ app/(ui)/dashboard/invoices/table.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
-import { UpdateInvoice, DeleteInvoice } from '@/app/ui/dashboard/invoices/buttons';
-import InvoiceStatus from '@/app/ui/dashboard/invoices/status';
+import { UpdateInvoice, DeleteInvoice } from '@/app/(ui)/dashboard/invoices/buttons';
+import InvoiceStatus from '@/app/(ui)/dashboard/invoices/status';
 import { formatDateToLocal, formatCurrency } from '@/app/lib/utils';
 import { fetchFilteredInvoices } from '@/app/lib/data';
