+++ app/(ui)/dashboard/sidenav.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
-import NavLinks from '@/app/ui/dashboard/nav-links';
-import AcmeLogo from '@/app/ui/acme-logo';
+import NavLinks from '@/app/(ui)/dashboard/nav-links';
+import AcmeLogo from '@/app/(ui)/acme-logo';
 import { PowerIcon } from '@heroicons/react/24/outline';
 
 export default function SideNav() {
