+++ app/page.tsx
@@ -1,8 +1,8 @@
-import AcmeLogo from '@/app/ui/acme-logo';
+import AcmeLogo from '@/app/(ui)/acme-logo';
 import { ArrowRightIcon } from '@heroicons/react/24/outline';
 import Link from 'next/link';
 import Image from 'next/image';
-import { lusitana } from '@/app/ui/fonts';
+import { lusitana } from '@/app/(ui)/fonts';
 
 export default function Page() {
   return (
